pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dllm-reason"
version = "0.1.0"
description = "DAG-guided discrete diffusion language models for reasoning"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "accelerate>=0.25.0",
    "networkx>=3.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "einops>=0.7.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
library = [
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
dllm-train = "dllm_reason.cli:train"
dllm-eval = "dllm_reason.cli:evaluate"
dllm-eval-dags = "dllm_reason.cli:eval_dags"
dllm-search = "dllm_reason.cli:search_dag"
dllm-viz = "dllm_reason.cli:visualize_dag"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
